+++ src/Grant/RefreshTokenGrant.php
@@ -61,6 +61,8 @@ class RefreshTokenGrant extends AbstractGrant
             }
         }
 
+        $scopes = $this->scopeRepository->finalizeScopes($scopes, $this->getIdentifier(), $client);
+
         // Expire old tokens
         $this->accessTokenRepository->revokeAccessToken($oldRefreshToken['access_token_id']);
         $this->refreshTokenRepository->revokeRefreshToken($oldRefreshToken['refresh_token_id']);